requirements.txt
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-apimanagement>=4.0.0
azure-mgmt-web>=7.0.0
azure-mgmt-containerinstance>=10.0.0
azure-mgmt-monitor>=6.0.0
azure-functions>=1.18.0
python-dotenv>=1.0.0
requests>=2.31.0
pyjwt>=2.8.0
cryptography>=41.0.0
jupyter>=1.0.0
notebook>=7.0.0
